requirements.txt
pytesseract~=0.3.10
PySide6~=6.3.0
openpyxl~=3.0.10
opencv-python~=4.6.0.66
Pillow~=9.1.0
PyPDF2~=1.28.4
pdf2image~=1.16.0